+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- 'SentryFileManager+Test.h' file not found (#3950)
+
 ## 8.25.1
 
 ### Features